Bounding box fixed definately - elastic collisions ok

+++ test/makepngfromps.sh
@@ -3,14 +3,21 @@
 
 mkdir -p framespng
 
-for((i=0;i<=10;i++))
+psfiles="analyse/frames/*.ps"
+#for file in "${folderps}/*.ps"
+for file in $psfiles
+#for((i=0;i<=10;i++))
 do
-	i=$(printf "%05d\n" "$i")
-	fileps="analyse/frames/${i}frame.ps"
-	filepng="${i}frame.png"
-	echo "$fileps -> $filepng"
-	ps2pdf -dPDFSETTINGS=/prepress -dEPSCrop $fileps tmp.pdf
-	gs -sDEVICE=pngalpha -r3000 -o $filepng tmp.pdf
+	[ -e "$file" ] || continue
+	echo $file
+	filepng="`echo "$file" | cut -d'.' -f1`.png"
+	echo $filepng
+#	i=$(printf "%05d\n" "$i")
+#	fileps="analyse/frames/${i}frame.ps"
+#	filepng="${i}frame.png"
+#	echo "$fileps -> $filepng"
+	ps2pdf -dPDFSETTINGS=/prepress -dEPSCrop $file tmp.pdf
+	gs -sDEVICE=pngalpha -r2000 -o $filepng tmp.pdf
 	mv $filepng framespng
 done
